requirements.txt
streamlit
google-generativeai
feedparser
requests
beautifulsoup4

plotly
